화면에서 사용자 삭제에 확인창을 추가하라
화면에서 사용자 탈퇴버튼을 누르면 탈퇴하겠냐는 확인창을 추가합니다

+++ app/src/main/resources/static/js/app/index.js
@@ -217,6 +217,10 @@ var main = {
 
     deleteUser : function () {
         var id = $('#id').val();
+        var result = confirm("정말로 탈퇴하시겠습니까?");
+        if(!result) {
+            return;
+        }
 
         $.ajax({
             type: 'DELETE',
@@ -224,7 +228,7 @@ var main = {
             dataType: 'json',
             contentType:'application/json; charset=utf-8',
         }).done(function() {
-            alert('사용자 삭제가 완료되었습니다.');
+            alert('사용자 탈퇴가 완료되었습니다.');
             window.location.href = '/';
         }).fail(function (request) {
             alert(request.responseText);
